+++ app/src/main/java/com/waz/zclient/pages/main/MainTabletFragment.java
@@ -43,7 +43,6 @@ import com.waz.zclient.core.stores.inappnotification.KnockingEvent;
 import com.waz.zclient.fragments.ImageFragment;
 import com.waz.zclient.pages.BaseFragment;
 import com.waz.zclient.pages.main.backgroundmain.views.BackgroundFrameLayout;
-import com.waz.zclient.pages.main.conversation.SingleImageFragment;
 import com.waz.zclient.pages.main.conversation.SingleImageUserFragment;
 import com.waz.zclient.pages.main.conversation.VideoPlayerFragment;
 import com.waz.zclient.pages.main.conversationlist.ConfirmationFragment;
@@ -57,7 +56,7 @@ public class MainTabletFragment extends BaseFragment<MainTabletFragment.Containe
                                                                                    OnBackPressedListener,
                                                                                    RootFragment.Container,
                                                                                    SingleImageObserver,
-                                                                                   SingleImageFragment.Container,
+                                                                                   SingleImageUserFragment.Container,
                                                                                    ConfirmationObserver,
                                                                                    AccentColorObserver,
                                                                                    ConfirmationFragment.Container,
@@ -161,8 +160,8 @@ public class MainTabletFragment extends BaseFragment<MainTabletFragment.Containe
 
         if (getChildFragmentManager().getBackStackEntryCount() > 0) {
             Fragment topFragment = getChildFragmentManager().findFragmentByTag(getChildFragmentManager().getBackStackEntryAt(0).getName());
-            if (topFragment instanceof SingleImageFragment) {
-                return ((SingleImageFragment) topFragment).onBackPressed();
+            if (topFragment instanceof SingleImageUserFragment) {
+                return ((SingleImageUserFragment) topFragment).onBackPressed();
             } else if (topFragment instanceof ConfirmationFragment) {
                 return ((ConfirmationFragment) topFragment).onBackPressed();
             }
